+++ site/src/pages/DeploymentSettingsPage/OverviewPage/OverviewPage.tsx
@@ -1,10 +1,5 @@
 import { deploymentDAUs } from "api/queries/deployment";
-import {
-	availableExperiments,
-	experiments,
-	isKnownExperiment,
-} from "api/queries/experiments";
-import { useEmbeddedMetadata } from "hooks/useEmbeddedMetadata";
+import { availableExperiments } from "api/queries/experiments";
 import { useDeploymentConfig } from "modules/management/DeploymentConfigProvider";
 import type { FC } from "react";
 import { useQuery } from "react-query";
@@ -15,14 +10,7 @@ const OverviewPage: FC = () => {
 	const { deploymentConfig } = useDeploymentConfig();
 	const safeExperimentsQuery = useQuery(availableExperiments());
 
-	const { metadata } = useEmbeddedMetadata();
-	const enabledExperimentsQuery = useQuery(experiments(metadata.experiments));
-
 	const safeExperiments = safeExperimentsQuery.data?.safe ?? [];
-	const invalidExperiments =
-		enabledExperimentsQuery.data?.filter((exp) => {
-			return !isKnownExperiment(exp);
-		}) ?? [];
 
 	const { data: dailyActiveUsers } = useQuery(deploymentDAUs());
 
@@ -33,7 +21,6 @@ const OverviewPage: FC = () => {
 			<OverviewPageView
 				deploymentOptions={deploymentConfig.options}
 				dailyActiveUsers={dailyActiveUsers}
-				invalidExperiments={invalidExperiments}
 				safeExperiments={safeExperiments}
 			/>
 		</>
