+++ site/src/pages/DeploymentSettingsPage/OverviewPage/OverviewPageView.stories.tsx
@@ -36,7 +36,6 @@ const meta: Meta<typeof OverviewPageView> = {
 			},
 		],
 		dailyActiveUsers: MockDeploymentDAUResponse,
-		invalidExperiments: [],
 		safeExperiments: [],
 	},
 	parameters: {
@@ -83,42 +82,5 @@ export const allExperimentsEnabled: Story = {
 			},
 		],
 		safeExperiments: ["example"],
-		invalidExperiments: [],
-	},
-};
-
-export const invalidExperimentsEnabled: Story = {
-	args: {
-		deploymentOptions: [
-			{
-				name: "Access URL",
-				description:
-					"The URL that users will use to access the Coder deployment.",
-				flag: "access-url",
-				flag_shorthand: "",
-				value: "https://dev.coder.com",
-				hidden: false,
-			},
-			{
-				name: "Wildcard Access URL",
-				description:
-					'Specifies the wildcard hostname to use for workspace applications in the form "*.example.com".',
-				flag: "wildcard-access-url",
-				flag_shorthand: "",
-				value: "*--apps.dev.coder.com",
-				hidden: false,
-			},
-			{
-				name: "Experiments",
-				description:
-					"Enable one or more experiments. These are not ready for production. Separate multiple experiments with commas, or enter '*' to opt-in to all available experiments.",
-				flag: "experiments",
-				value: ["invalid", "*"],
-				flag_shorthand: "",
-				hidden: false,
-			},
-		],
-		safeExperiments: ["example"],
-		invalidExperiments: ["invalid"],
 	},
 };
